Move logic under feature.experimental.betterMatchTypesExtractors
This way we can merge this PR without waiting for the SIP committee to approve
it.
