+++ tests/neg/mt-deskolemize-2.scala
@@ -2,56 +2,59 @@
 
 trait Expr:
   type Value
-
 object Expr:
   type Of[V] = Expr { type Value = V }
   type ExtractValue[F <: Expr] = F match
     case Expr.Of[v] => v
 import Expr.ExtractValue
 
-class Prim extends Expr:
-  type Value = Alias
-  type Alias = BigInt
+class SimpleLoop1 extends Expr:
+  type Value = ExtractValue[SimpleLoop2]
 
-class VecExpr[E <: Expr] extends Expr:
-  type Value = Vector[ExtractValue[E]]
+class SimpleLoop2 extends Expr:
+  type Value = ExtractValue[SimpleLoop1]
+
+object Test1:
+  val x: ExtractValue[SimpleLoop1] = 1 // error
 
 trait Description:
   type Elem <: Tuple
+
+class PrimBroken extends Expr:
+  type Value = Alias
+  type Alias = Value // error
+
+class Prim extends Expr:
+  type Value = BigInt
+
+class VecExpr[E <: Expr] extends Expr:
+  type Value = Vector[ExtractValue[E]]
 
 trait ProdExpr extends Expr:
   val description: Description
   type Value = Tuple.Map[description.Elem, [X] =>> ExtractValue[X & Expr]]
 
+
 class MyExpr1 extends ProdExpr:
   final val description = new Description:
-    type Elem = (VecExpr[Prim], Prim)
+    type Elem = (VecExpr[Prim], MyExpr2)
 
 class MyExpr2 extends ProdExpr:
   final val description = new Description:
     type Elem = (VecExpr[VecExpr[MyExpr1]], Prim)
-
-trait ProdExprAlt[T <: Tuple] extends Expr:
-  type Value = Tuple.Map[T, [X] =>> ExtractValue[X & Expr]]
-
-class MyExpr3 extends ProdExprAlt[(Prim, VecExpr[Prim], Prim)]
 
 trait Constable[E <: Expr]:
   def lit(v: ExtractValue[E]): E
 object Constable:
   given [E <: Expr]: Constable[E] = ???
 
-object Test:
+object Test2:
   def fromLiteral[E <: Expr : Constable](v: ExtractValue[E]): E =
     summon[Constable[E]].lit(v)
-  val a: Prim = fromLiteral(1)
-  val b: VecExpr[Prim] = fromLiteral(Vector(1))
-  val c: MyExpr1 = fromLiteral((Vector(1), 1))
-  val d: MyExpr2 = fromLiteral(Vector(Vector((Vector(1), 1))), 2)
-  val e: MyExpr3 = fromLiteral((1, Vector(1), 1))
-  val f: ProdExprAlt[(MyExpr1, VecExpr[MyExpr3])] = fromLiteral:
-    (
-      (Vector(1), 1),
-      Vector((1, Vector(1), 1), (2, Vector(1), 2))
-    )
-  val g: Expr { type Alias = Int; type Value = Alias } = fromLiteral(1)
+  val x0: ExtractValue[Prim] = "" // error
+  val x1: ExtractValue[PrimBroken] = 1 // error
+
+  val foo: MyExpr2 = new MyExpr2
+  val v: foo.Value = (Vector(Vector()), 1) // error: Recursion limit exceeded
+  val c: MyExpr2 = fromLiteral:
+    (Vector(Vector()), 1) // error: Recursion limit exceeded
